Fix heap corruption in StelemRef on negative index
StelemRef with a negative index would result in a fun heap corruption. The `IList_Generic_ItemSet_NegativeIndex_ThrowsException` test in System.Runtime.Tests was leaving us with a corrupted heap.

+++ src/coreclr/nativeaot/Runtime.Base/src/System/Runtime/TypeCast.cs
@@ -725,7 +725,7 @@ namespace System.Runtime
             // This is supported only on arrays
             Debug.Assert(array.EEType->IsArray, "first argument must be an array");
 
-            if (index >= array.Length)
+            if ((uint)index >= (uint)array.Length)
             {
                 throw array.EEType->GetClasslibException(ExceptionIDs.IndexOutOfRange);
             }
